move code around a bit
